Moved and splitted mappoint series into subfolder.
