+++ js/Series/MapPoint/MapPointPoint.js
@@ -1,0 +1,67 @@
+/* *
+ *
+ *  (c) 2010-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+'use strict';
+var __extends = (this && this.__extends) || (function () {
+    var extendStatics = function (d, b) {
+        extendStatics = Object.setPrototypeOf ||
+            ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
+            function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
+        return extendStatics(d, b);
+    };
+    return function (d, b) {
+        extendStatics(d, b);
+        function __() { this.constructor = d; }
+        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
+    };
+})();
+import BaseSeries from '../../Core/Series/Series.js';
+var ScatterSeries = BaseSeries.seriesTypes.scatter;
+import U from '../../Core/Utilities.js';
+var merge = U.merge;
+/* *
+ *
+ *  Class
+ *
+ * */
+var MapPointPoint = /** @class */ (function (_super) {
+    __extends(MapPointPoint, _super);
+    function MapPointPoint() {
+        /* *
+         *
+         *  Properties
+         *
+         * */
+        var _this = _super !== null && _super.apply(this, arguments) || this;
+        _this.options = void 0;
+        _this.series = void 0;
+        return _this;
+        /* eslint-enable valid-jsdoc */
+    }
+    /* *
+     *
+     *  Functions
+     *
+     * */
+    /* eslint-disable valid-jsdoc */
+    MapPointPoint.prototype.applyOptions = function (options, x) {
+        var mergedOptions = (typeof options.lat !== 'undefined' &&
+            typeof options.lon !== 'undefined' ?
+            merge(options, this.series.chart.fromLatLonToPoint(options)) :
+            options);
+        return _super.prototype.applyOptions.call(this, mergedOptions, x);
+    };
+    return MapPointPoint;
+}(ScatterSeries.prototype.pointClass));
+/* *
+ *
+ *  Default Export
+ *
+ * */
+export default MapPointPoint;
