+++ js/masters/modules/map.src.js
@@ -17,7 +17,7 @@ import '../../Maps/MapNavigation.js';
 import '../../Maps/MapPointer.js';
 import '../../Series/Map/MapSeries.js';
 import '../../Series/MapLineSeries.js';
-import '../../Series/MapPointSeries.js';
+import '../../Series/MapPoint/MapPointSeries.js';
 import '../../Series/MapBubbleSeries.js';
 import '../../Series/HeatmapSeries.js';
 import '../../Extensions/GeoJSON.js';
