+++ ts/Series/MapPoint/MapPointPointOptions.d.ts
@@ -1,0 +1,32 @@
+/* *
+ *
+ *  (c) 2010-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type ScatterPointOptions from '../Scatter/ScatterPointOptions';
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+export interface MapPointPointOptions extends ScatterPointOptions {
+    lat?: number;
+    lon?: number;
+    x?: number;
+    y?: (number|null);
+}
+
+export default MapPointPointOptions;
